refactor: 기존 카카오맵 호출하는 로직 제거 및 리팩토링
